4.23.22
the original idea for index.js is getting overly complicated, so ive added index2.js to build out the core functionality of the application for now, and if i decide to add in all the other features i would like to have then the original index can be revisited. otherwise it will likely just be deleted.