4.25.22
built prompts and welcome functions, tested to make sure employee objects are being generated correctly and are puushed to the output array. Next step is linking those objects to the classes that will be used to make the HTML page. I would still like to add in the ability to build different teams and go back and edit them but that will be a down the road feature add-in